requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-rest-auth==0.9.5
djangorestframework==3.12.4
Faker==8.10.3
idna==3.2
model-mommy==2.0.0
oauthlib==3.1.1
Pillow==8.3.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
